Integrating the IntentDeclaration into the IntentInfo visitor
